--- a/be/src/column/binary_column.cpp
+++ b/be/src/column/binary_column.cpp
@@ -67,10 +67,10 @@ void BinaryColumnBase<T>::append_selective(const Column& src, const uint32_t* in
 
     auto* dest_bytes = _bytes.data();
     for (size_t i = 0; i < size; i++) {
-        uint32_t row_idx = indexes[from + i];
-        T str_size = src_offsets[row_idx + 1] - src_offsets[row_idx];
-        strings::memcpy_inlined(dest_bytes + _offsets[cur_row_count + i], src_bytes.data() + src_offsets[row_idx],
-                                str_size);
+        T src_offset = src_offsets[from + i];
+        T src_length = src_offsets[from + i + 1] - src_offset;
+        T dst_offset = _offsets[cur_row_count + i];
+        strings::memcpy_inlined(dest_bytes + dst_offset, src_bytes.data() + src_offset, src_length);
     }
 
     _slices_cache = false;