@mofeiatwork 2022-10-31
run starrocks_admit_test